Fix typo in Technical Gov Timeline Recommendation
There was wording in the Technical Governance Recommendation to the
Board approved today that might be slightly confusing due to life
editing during the meetings. This patch proposes one possible
improvement to make things clearer to the Board.

I do not believe this patch substantively changes the recommendation
at all.

Change-Id: Ia58000f04a6bfea6d525257220d3beeeee6cddb9
Signed-off-by: Ed Warnicke <hagbard@gmail.com>
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>